pyproject.toml
[tool.poetry]
name = "toktik-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
boto3 = "^1.28.63"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
fastapi = "^0.103.2"
fastapi-jwt = "^0.1.12"
pydantic = "^2.4.2"
pika = "^1.3.2"
sqlalchemy = "^2.0.22"
pymysql = "^1.1.0"
python-socketio = "^5.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
